src/data_prep.py
"""
Data preparation script for NLP multi-type classification project.

This script:
1. Reads raw data files organized by families
2. Validates and transforms into single-sentence samples
3. Performs family-aware train/val/test split (no leakage)
4. Writes processed CSVs to data/processed/
5. Generates validation report

Usage:
    python src/data_prep.py --config configs/data_config.yaml
"""

import argparse
import yaml
from pathlib import Path
from typing import Dict, List, Tuple, Any
import sys

# TODO: Add imports when implementing
# import pandas as pd
# import numpy as np
# from sklearn.model_selection import GroupShuffleSplit

from constants import (
    LABELS,
    LABEL2ID,
    REQUIRED_COLUMNS,
    DEFAULT_SEED,
    SPLIT_RATIOS,
    GROUP_KEY,
    MAX_TEXT_LENGTH_CHARS,
    MIN_CLASS_PROPORTION,
    TRAIN_FILENAME,
    VAL_FILENAME,
    TEST_FILENAME,
)


def load_config(config_path: str) -> Dict[str, Any]:
    """
    Load data configuration from YAML file.
    
    Args:
        config_path: Path to data_config.yaml
        
    Returns:
        Configuration dictionary
    """
    # TODO: Implement YAML loading and validation
    pass


def load_raw_data(raw_dir: Path) -> Any:
    """
    Load raw data files from data/raw directory.
    
    Expected format: JSON or CSV files with family-based organization.
    Each row should have: family_id, type1, type2, type3, type4.
    
    Args:
        raw_dir: Path to raw data directory
        
    Returns:
        DataFrame or list of family records
        
    TODO: Implement based on actual raw data format
    """
    # TODO: Implement raw data loading
    # Handle multiple file formats (JSON, CSV, etc.)
    # Validate that required fields exist
    pass


def build_sentence_samples(raw_data: Any) -> Any:
    """
    Transform family-based raw data into single-sentence samples.
    
    For each family:
    - Extract non-null values from type1, type2, type3, type4
    - Create one row per sentence with columns: family_id, text, label
    - Compute text_len_char and text_len_word (optional)
    
    Args:
        raw_data: Raw data loaded from files
        
    Returns:
        DataFrame with columns: family_id, text, label, text_len_char, text_len_word
        
    TODO: Implement transformation logic
    """
    # TODO: Implement
    # Iterate through families
    # For each type field (type1..type4), create a ProcessedRow
    # Apply label mapping from constants.LABEL2ID
    pass


def preprocess_text(text: str, config: Dict[str, Any]) -> str:
    """
    Apply preprocessing rules to text.
    
    Based on preprocessing config:
    - strip_whitespace: Remove leading/trailing whitespace
    - normalize_unicode: Apply unicode normalization (NFC)
    - Additional cleaning as needed
    
    Args:
        text: Input text
        config: Preprocessing configuration from data_config.yaml
        
    Returns:
        Cleaned text
        
    TODO: Implement preprocessing
    """
    # TODO: Implement
    pass


def deduplicate_samples(df: Any) -> Tuple[Any, int]:
    """
    Remove exact duplicate texts within the dataset.
    
    Strategy:
    - Drop exact duplicate text strings
    - Log count of duplicates removed
    - Flag cross-split duplicates if found (potential leakage)
    
    Args:
        df: DataFrame with samples
        
    Returns:
        Tuple of (deduplicated DataFrame, count of removed duplicates)
        
    TODO: Implement deduplication
    """
    # TODO: Implement
    pass


def family_aware_split(
    df: Any, 
    group_key: str,
    train_ratio: float,
    val_ratio: float,
    test_ratio: float,
    seed: int
) -> Tuple[Any, Any, Any]:
    """
    Perform family-aware train/val/test split.
    
    Critical: All samples from the same family_id must go to the same split.
    Use GroupShuffleSplit or similar to ensure no leakage.
    
    Args:
        df: DataFrame with all samples
        group_key: Column name for grouping (typically 'family_id')
        train_ratio: Proportion for training set
        val_ratio: Proportion for validation set
        test_ratio: Proportion for test set
        seed: Random seed for reproducibility
        
    Returns:
        Tuple of (train_df, val_df, test_df)
        
    TODO: Implement group-aware splitting
    """
    # TODO: Implement
    # Use sklearn.model_selection.GroupShuffleSplit or custom logic
    # First split into train and temp (val+test)
    # Then split temp into val and test
    # Verify no group appears in multiple splits
    pass


def validate_splits(train_df: Any, val_df: Any, test_df: Any) -> Dict[str, Any]:
    """
    Run validation checks on splits.
    
    Checks:
    1. No family_id overlap between splits (leakage check)
    2. All labels are valid
    3. Class distribution in each split
    4. Warn if class imbalance > threshold
    5. Text length statistics
    
    Args:
        train_df: Training DataFrame
        val_df: Validation DataFrame
        test_df: Test DataFrame
        
    Returns:
        Dictionary with validation results and summary statistics
        
    TODO: Implement validation logic
    """
    # TODO: Implement
    # Check for family_id overlaps
    # Compute class ratios
    # Generate summary statistics
    pass


def write_processed_data(train_df: Any, val_df: Any, test_df: Any, output_dir: Path):
    """
    Write processed DataFrames to CSV files.
    
    Output files:
    - train_4class.csv
    - val_4class.csv
    - test_4class.csv
    
    Args:
        train_df: Training samples
        val_df: Validation samples
        test_df: Test samples
        output_dir: Directory to write CSVs (data/processed/)
        
    TODO: Implement CSV writing
    """
    # TODO: Implement
    # Write with consistent column order from constants.COLUMN_NAMES
    # Use UTF-8 encoding
    # Add index=False
    pass


def generate_summary_report(validation_results: Dict[str, Any]) -> str:
    """
    Generate human-readable summary report of data preparation.
    
    Include:
    - Total samples per split
    - Class distribution per split
    - Leakage check status (PASS/FAIL)
    - Text length statistics
    - Any warnings (imbalance, outliers, etc.)
    
    Args:
        validation_results: Results from validate_splits()
        
    Returns:
        Formatted report string
        
    TODO: Implement report generation
    """
    # TODO: Implement
    pass


def main():
    """
    Main data preparation pipeline.
    
    Steps:
    1. Load configuration
    2. Load raw data
    3. Build single-sentence samples
    4. Preprocess and deduplicate
    5. Perform family-aware split
    6. Validate splits
    7. Write processed CSVs
    8. Print summary report
    """
    parser = argparse.ArgumentParser(description="Prepare data for multi-type classification")
    parser.add_argument(
        "--config",
        type=str,
        default="configs/data_config.yaml",
        help="Path to data configuration file"
    )
    args = parser.parse_args()
    
    print("=" * 60)
    print("NLP Multi-Type Classification: Data Preparation")
    print("=" * 60)
    
    # TODO: Implement pipeline steps
    # Step 1: Load config
    # config = load_config(args.config)
    
    # Step 2: Load raw data
    # raw_data = load_raw_data(Path(config['raw_dir']))
    
    # Step 3: Build samples
    # samples_df = build_sentence_samples(raw_data)
    
    # Step 4: Preprocess
    # samples_df['text'] = samples_df['text'].apply(lambda x: preprocess_text(x, config['preprocessing']))
    
    # Step 5: Deduplicate
    # samples_df, dup_count = deduplicate_samples(samples_df)
    
    # Step 6: Split
    # train_df, val_df, test_df = family_aware_split(...)
    
    # Step 7: Validate
    # validation_results = validate_splits(train_df, val_df, test_df)
    
    # Step 8: Write
    # write_processed_data(train_df, val_df, test_df, Path(config['processed_dir']))
    
    # Step 9: Report
    # report = generate_summary_report(validation_results)
    # print(report)
    
    print("\n[TODO] Data preparation pipeline not yet implemented.")
    print("This script defines the structure and will be implemented in the next phase.")
    
    return 0


if __name__ == "__main__":
    sys.exit(main())

