requirements.txt
# NLP Multi-Type Classification Project - Requirements
# Python 3.8+ required

# ============================================================
# Core Data Science Libraries
# ============================================================
pandas>=1.5.0,<3.0.0
numpy>=1.23.0,<2.0.0
scipy>=1.10.0,<2.0.0

# ============================================================
# Machine Learning - Classical
# ============================================================
scikit-learn>=1.2.0,<2.0.0
xgboost>=1.7.0,<3.0.0

# ============================================================
# Machine Learning - Deep Learning
# ============================================================
torch>=2.0.0,<3.0.0
transformers>=4.30.0,<5.0.0
datasets>=2.12.0,<3.0.0
accelerate>=0.20.0,<1.0.0

# ============================================================
# Configuration and I/O
# ============================================================
pyyaml>=6.0,<7.0
python-dotenv>=1.0.0,<2.0.0

# ============================================================
# Visualization
# ============================================================
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# ============================================================
# Jupyter and Interactive Development
# ============================================================
jupyter>=1.0.0,<2.0.0
ipykernel>=6.23.0,<7.0.0
notebook>=6.5.0,<8.0.0

# ============================================================
# Data Validation (Optional)
# ============================================================
pydantic>=2.0.0,<3.0.0

# ============================================================
# Utilities
# ============================================================
tqdm>=4.65.0,<5.0.0
requests>=2.28.0,<3.0.0

# ============================================================
# AWS Integration (Optional - for AWS deployment)
# ============================================================
# Uncomment these when deploying to AWS:
# boto3>=1.26.0,<2.0.0
# sagemaker>=2.150.0,<3.0.0
# s3fs>=2023.5.0,<2024.0.0

# ============================================================
# Testing (Optional - for development)
# ============================================================
# pytest>=7.3.0,<8.0.0
# pytest-cov>=4.1.0,<5.0.0

# ============================================================
# Code Quality (Optional - for development)
# ============================================================
# black>=23.3.0,<24.0.0
# flake8>=6.0.0,<7.0.0
# mypy>=1.3.0,<2.0.0

# ============================================================
# Experiment Tracking (Optional)
# ============================================================
# mlflow>=2.3.0,<3.0.0
# wandb>=0.15.0,<1.0.0
# tensorboard>=2.13.0,<3.0.0

